pyproject.toml
[tool.poetry]
name = "cac-24262-fullstack-python"
version = "0.1.0"
description = ""
authors = ["Lucas Lucyk <lucaslucyk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
fastapi = "^0.111.0"
pydantic = "^2.8.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
